packages/sheratan-orchestrator/requirements.txt
python-dotenv==1.0.0
aiohttp==3.9.1
beautifulsoup4==4.12.2
